Refactoring, because it was bugging me.
