+++ lib/galaxy/datatypes/sequence.py
@@ -15,9 +15,9 @@ from six import PY3
 from galaxy import util
 from galaxy.datatypes import metadata
 from galaxy.util.checkers import is_gzip
+from galaxy.util.image_util import check_image_type
 from galaxy.datatypes.metadata import MetadataElement
 from galaxy.datatypes.sniff import get_headers
-from galaxy.datatypes.util.image_util import check_image_type
 from galaxy.util import nice_size
 from . import data
