+++ lib/galaxy/util/checkers.py
@@ -1,20 +1,14 @@
 import bz2
 import gzip
-import imghdr
 import re
 import zipfile
 
 from six import StringIO
 
 from galaxy import util
+from galaxy.util.image_util import image_type
 
 HTML_CHECK_LINES = 100
-
-
-def check_image( file_path ):
-    if imghdr.what( file_path ) is not None:
-        return True
-    return False
 
 
 def check_html( file_path, chunk=None ):
@@ -124,6 +118,13 @@ def is_bz2( file_path ):
 def is_gzip( file_path ):
     is_gzipped, is_valid = check_gzip( file_path )
     return is_gzipped
+
+
+def check_image( file_path ):
+    """ Simple wrapper around image_type to yield a True/False verdict """
+    if image_type( file_path ):
+        return True
+    return False
 
 
 __all__ = [
