+++ lib/galaxy/util/image_util.py
